Unify v7/v8 exercisers
Also pass in queue name as a parameter to StartBus to help with test assertions
